% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{addRowToBeta}
\alias{addRowToBeta}
\title{split group q of provided beta randomly into two}
\usage{
addRowToBeta(beta, q)
}
\arguments{
\item{beta}{provided beta}

\item{q}{index of group to split}
}
\value{
new beta
}
\description{
split group q of provided beta randomly into two
}
