% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{analyze_microbiome_metabolite_network}
\alias{analyze_microbiome_metabolite_network}
\title{Microbiome-Metabolite Network Analysis Wrapper}
\usage{
analyze_microbiome_metabolite_network(
  X,
  Y,
  method,
  lambda = NULL,
  alpha = NULL,
  zscore_method = NULL,
  sbm_model = NULL,
  sbm_params = NULL,
  nb_cores = NULL
)
}
\arguments{
\item{X}{A p x n matrix of microbiome abundances (p features, n samples)}

\item{Y}{A q x n matrix of metabolite abundances (q features, n samples)}

\item{method}{Character string specifying the method to use.
One of: 
  \describe{
    \item{"corr"}{Marginal correlation}
    \item{"pcorr"}{Partial correlation (e.g., graphical lasso)}
    \item{"corr_test"}{Marginal correlation after accounting for multiple comparisons}
  }}

\item{lambda}{A sequence of decreasing positive numbers for regularization (used in pcorr method)}

\item{alpha}{Significance level for graph inference (default: 0.05)}

\item{zscore_method}{Character string specifying z-score normalization for corr_test method
One of: 
  \describe{
    \item{"pearson+fisher"}{Uses Pearson correlation, followed by Fisher transform}
    \item{"spearman+fisher"}{Uses Spearman correlation, followed by Fisher transform}
    \item{"CL"}{Uses the method from Cai and Liu}
  }}

\item{sbm_model}{Character string specifying the model type for corr_test method 
One of: 
  \describe{
    \item{"Gauss0"}{the mean of the null distribution is set to 0}
    \item{"Gauss01"}{the null distribution is set to N(0,1)}
    \item{"Gauss02distr"}{the mean of the null distribution is set to 0 and the alternative distribution is a single Gaussian distribution, i.e. the block memberships of the nodes do not influence on the alternative distribution}
  }}

\item{sbm_params}{List of parameters for corr_test method (Q1, Q2, explor)}

\item{nb_cores}{Number of cores for parallel processing for corr_test method}
}
\value{
A named list with method-dependent result

\strong{If \code{method = "corr"}}:
\describe{
  \item{result}{Estimated correlation matrix (p+q) x (p+q)}
}

\strong{If \code{method = "pcorr"}}:
\describe{
  \item{result}{Estimated partial correlation matrix (p+q) x (p+q)}
}

\strong{If \code{method = "corr_test"}}:
\describe{
  \item{qvalues}{(p x q) matrix of FDR-adjusted p-values for each cross-feature test}
}
}
\description{
Microbiome-Metabolite Network Analysis Wrapper
}
