% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{betaUp}
\alias{betaUp}
\title{Create new values of beta by splitting groups of provided beta}
\usage{
betaUp(beta, nbOfSplits = 1)
}
\arguments{
\item{beta}{a matrix consisting of soft node clustering for row or column}

\item{nbOfSplits}{number of required splits of blocks}
}
\value{
a list of length nbOfSplits (at most) of initial points for beta
}
\description{
Create nbOfSplits (or all) new values of beta by splitting nbOfSplits (or all) groups of provided beta
}
