% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm_biSBM.R
\name{betaUpdate}
\alias{betaUpdate}
\title{Compute one iteration to solve the fixed point equation in the VE-step}
\usage{
betaUpdate(beta.list, log.pi, log.1mpi, dataMatrix, VE, mstep, modelFamily)
}
\arguments{
\item{beta.list}{current value of beta in a list of two}

\item{log.pi}{value of log(w)}

\item{log.1mpi}{value of log(1-w)}

\item{dataMatrix}{observed dense adjacency matrix}

\item{VE}{list with variational parameters beta and rho}

\item{mstep}{list with current model parameters and additional auxiliary terms}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}
}
\value{
updated value of beta
}
\description{
Compute one iteration to solve the fixed point equation in the VE-step
}
