% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm.corr.R
\name{cggm.corr}
\alias{cggm.corr}
\title{Function to estimate the correlation matrix}
\usage{
cggm.corr(X)
}
\arguments{
\item{X}{A sample by variable data matrix}
}
\value{
The estimated correlation matrix
}
\description{
\code{cggm.corr} estimates the correlation between (potentially censored) continuous variables using the framework
of censored Gaussian graphical models.
}
\details{
When no variable is censored, \code{cggm.corr} estimates the correlation using Pearson's correlation coefficient.
When there are censored variables, a univariate Tobit model is first used to estimate the marginal distribution of
each variable. Pairwise correlations are estimated using maximum likelihood after plugging in the marginal mean
and standard deviation estimates.

The correlation matrix based on pairwise estimates is not necessarily positive definite. When this is the case,
the function \code{\link[Matrix]{nearPD}} is used to compute the nearest positive definite correlation matrix with a
tolerance 1e-04 for enforcing positive definiteness.
}
\examples{
library(MASS)
p <- 20
S <- diag(1,p)
for (j in 1:(p-1)){
   S[j,j+1] <- S[j+1,j] <- 0.5
}
X <- mvrnorm(n=100, mu=rep(0,p), Sigma=S)
X_cens <- apply(X,2,function(a) ifelse(a>1,a,0))
S_hat <- cggm.corr(X_cens)

}
\references{
Ma, Jing. Joint Microbial and Metabolomic Network Estimation with the Censored Gaussian Graphical Model. Technical Report. 2019.
}
