% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm.precision.R
\name{cggm.precision}
\alias{cggm.precision}
\title{Function to estimate the precision matrix based on the censored Gaussian graphical model}
\usage{
cggm.precision(X, lambda, method = c("clime", "glasso"))
}
\arguments{
\item{X}{A sample by variable (\code{n} by \code{p}) data matrix}

\item{lambda}{A sequence of decreasing positive numbers to control the regularization.}

\item{method}{Method for estimating the precision matrix. Could be either \code{clime} or \code{glasso}.}
}
\value{
A list with items
\item{path}{A list of \code{p} by \code{p} adjacency matrices of estimated graphs as a graph path corresponding to \code{lambda}}
\item{icov}{A list of \code{p} by \code{p} inverse covariance matrices corresponding to \code{lambda}}
\item{cov}{A list of \code{p} by \code{p} estimated covariance matrices corresponding to \code{lambda}}
\item{lambda}{The sequencing of decreasing regulatization parameters used in estimating the precision matrices}
\item{data}{The \code{n} by \code{p} data matrix from the input}
}
\description{
Function to estimate the precision matrix based on the censored Gaussian graphical model
}
