% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{classInd}
\alias{classInd}
\title{convert a clustering into a 0-1-matrix}
\usage{
classInd(cl, nbClusters)
}
\arguments{
\item{cl}{cluster in vector form}

\item{nbClusters}{number of clusters}
}
\value{
a 0-1-matrix encoding the clustering
}
\description{
convert a clustering into a 0-1-matrix
}
