% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{convertGroupPair}
\alias{convertGroupPair}
\title{transform a pair of block identifiers (q,l) into an identifying integer}
\usage{
convertGroupPair(q, l, Q1, Q2)
}
\arguments{
\item{q}{indicator of a latent block}

\item{l}{indicator of a latent block; can be a vector}

\item{Q1}{number of latent blocks on side one}

\item{Q2}{number of latent blocks on side two}

\item{directed}{indicates if the graph is directed}
}
\description{
this is the inverse function of convertGroupPairIdentifierBipartite()
}
