% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{convertGroupPairIdentifier}
\alias{convertGroupPairIdentifier}
\title{takes a scalar index of a group pair (q,l) and returns the values q and l}
\usage{
convertGroupPairIdentifier(ind_ql, Q)
}
\arguments{
\item{ind_ql}{indicator for a pair of latent blocks}

\item{Q}{number of latent blocks}

\item{Q1}{number of latent blocks on side one}

\item{Q2}{number of latent blocks on side two
takes a scalar indice of a group pair (q,l) and returns the values q and l

this is the inverse function of convertGroupPair()}
}
\description{
this is the inverse function of convertGroupPair()
}
