% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{correctBeta}
\alias{correctBeta}
\title{corrects values of the variational parameters tau that are too close to the 0 or 1}
\usage{
correctBeta(tau)
}
\arguments{
\item{tau}{variational parameters}
}
\description{
corrects values of the variational parameters tau that are too close to the 0 or 1
}
