% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{getBetaql}
\alias{getBetaql}
\title{Evaluate beta_q_1*beta_l_2 in the noisy bipartite stochastic block model}
\usage{
getBetaql(q, l, beta)
}
\arguments{
\item{q}{indicator of a latent block on side one}

\item{l}{indicator of a latent block on side two}

\item{beta}{variational parameters in a list of length 2}
}
\description{
Evaluate beta_q_1*beta_l_2 in the noisy bipartite stochastic block model
}
