% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{getRho}
\alias{getRho}
\title{compute rho associated with given values of pi, nu0 and nu}
\usage{
getRho(Q1, Q2, pi, nu0, nu, dataMatrix, modelFamily)
}
\arguments{
\item{Q1}{number of latent blocks in the noisy bipartite stochastic block model on side one}

\item{Q2}{number of latent blocks in the noisy bipartite stochastic block model on side two}

\item{pi}{connectivity parameter in the noisy bipartite stochastic block model}

\item{nu0}{null parameter in the noisy bipartite stochastic block model}

\item{nu}{alternative parameter in the noisy bipartite stochastic block model}

\item{dataMatrix}{observed dense adjacency matrix}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}
}
\value{
a matrix of conditional probabilities of an edge given the node memberships of the interacting nodes
}
\description{
compute rho associated with given values of pi, nu0 and nu
}
