% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure_biSBM.R
\name{graphInferenceBH}
\alias{graphInferenceBH}
\title{graph inference via the Benjamini-Hochberg multiple testing procedure}
\usage{
graphInferenceBH(dataMatrix, alpha = 0.05)
}
\description{
graph inference via the Benjamini-Hochberg multiple testing procedure
}
