% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure_biSBM.R
\name{graphInferenceLIS}
\alias{graphInferenceLIS}
\title{graph inference via the multiple testing procedure in Sun and Cai (09)
which accounts for dependency via a latent HMM model
LIS is short for local index of significance}
\usage{
graphInferenceLIS(dataMatrix, ncomp = 2)
}
\arguments{
\item{ncomp}{number of component in the mixture model. Default is 2.}
}
\description{
graph inference via the multiple testing procedure in Sun and Cai (09)
which accounts for dependency via a latent HMM model
LIS is short for local index of significance
}
