% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure_biSBM.R
\name{graphInferenceSC}
\alias{graphInferenceSC}
\title{graph inference via the multiple testing procedure in Sun and Cai (07)}
\usage{
graphInferenceSC(dataMatrix, alpha = 0.05)
}
\description{
graph inference via the multiple testing procedure in Sun and Cai (07)
}
