% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{graphRegularization}
\alias{graphRegularization}
\title{regularization of bipartite graphs}
\usage{
graphRegularization(A, tau, weighted = TRUE)
}
\arguments{
\item{A}{adjacency matrix}

\item{tau}{regularization parameter. Default is tau=3.}
}
\value{
a vector containing a node clustering into K groups
}
\description{
regularization of bipartite graphs
}
