% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{initialBeta}
\alias{initialBeta}
\title{compute initial values for beta}
\usage{
initialBeta(Q1, Q2, dataMatrix, method = "spectral")
}
\arguments{
\item{Q1}{number of latent blocks on side one in the noisy bipartite stochastic block model}

\item{Q2}{number of latent blocks on side two in the noisy bipartite stochastic block model}

\item{dataMatrix}{observed dense adjacency matrix}

\item{method}{method used for initializing the latent block memberships}
}
\value{
a list of initial points for beta1 and beta2
}
\description{
returns a list of length nbOfBeta of initial points for beta using spectral clustering with absolute values, kmeans and random perturbations of these points
}
