% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{initialPoints}
\alias{initialPoints}
\title{compute a list of initial points for the VEM algorithm}
\usage{
initialPoints(Q1, Q2, dataMatrix, modelFamily, model, method = "kmeans")
}
\arguments{
\item{Q1}{number of latent blocks in the Rows}

\item{Q2}{number of latent blocks in the Cols}

\item{dataMatrix}{observed dense adjacency matrix}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}

\item{model}{Implemented models:
\describe{
  \item{\code{Gauss}}{all Gaussian parameters of the null and the alternative distributions are unknown ; this is the Gaussian model with maximum number of unknown parameters}
  \item{\code{Gauss0}}{compared to \code{Gauss}, the mean of the null distribution is set to 0}
  \item{\code{Gauss01}}{compared to \code{Gauss}, the null distribution is set to N(0,1)}
  \item{\code{GaussEqVar}}{compared to \code{Gauss}, all Gaussian variances (of both the null and the alternative) are supposed to be equal, but unknown}
  \item{\code{Gauss0EqVar}}{compared to \code{GaussEqVar}, the mean of the null distribution is set to 0}
  \item{\code{Gauss0Var1}}{compared to \code{Gauss}, all Gaussian variances are set to 1 and the null distribution is set to N(0,1)}
  \item{\code{Gauss02distr}}{the mean of the null distribution is set to 0 and the alternative distribution is a single Gaussian distribution, i.e. the block memberships of the nodes do not influence on the alternative distribution}
  \item{\code{Gauss2distr}}{the alternative distribution is a single Gaussian distribution, i.e. the block memberships of the nodes do not influence on the alternative distribution}
  \item{\code{GaussAffil}}{compared to \code{Gauss}, for the alternative distribution, there's a distribution for inter-group and another for intra-group interactions}
  \item{\code{Exp}}{the null and the alternatives are all exponential distributions (i.e. Gamma distributions with shape parameter equal to one)  with unknown scale parameters}
  \item{\code{ExpGamma}}{the null distribution is an unknown exponential, the alterantive distribution are Gamma distributions with unknown parameters}
}}

\item{method}{type of method used when initializing the row/column block memberships}

\item{nbOfBeta}{number of initializations for the latent block memberships}

\item{nbOfPointsPerBeta}{number of initializations of the latent binary graph
associated with each initial latent block memberships}
}
\value{
list of inital points of beta and rho of length nbOfBeta*nbOfPointsPerBeta
}
\description{
compute a list of initial points of beta and rho for the VEM algorithm
for a given number of blocks; returns nbOfBeta*nbOfPointsPerBeta inital points
}
