% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure_biSBM.R
\name{lvaluesNobiSBM}
\alias{lvaluesNobiSBM}
\title{compute conditional l-values in the noisy bipartite stochastic block model}
\usage{
lvaluesNobiSBM(dataMatrix, Z1, Z2, theta, modelFamily = "Gauss")
}
\arguments{
\item{dataMatrix}{observation matrix}

\item{Z1}{a node clustering for rows}

\item{Z2}{a node clustering for columns}

\item{theta}{list of parameters for a noisy bipartite stochastic block model}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}}
}
\value{
conditional l-values in the noisy stochastic block model
}
\description{
compute conditional l-values in the noisy bipartite stochastic block model
}
