% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclr.R
\name{mclr}
\alias{mclr}
\title{Function to transform microbial count data}
\usage{
mclr(Z, base = exp(1), eps = 0.1)
}
\arguments{
\item{Z}{A sample by species count matrix}

\item{base}{The base of logarithm transformation. The default value is \code{e}.}

\item{eps}{A small constant added to adjust negative values after centered log-ratio transformation. See details.}
}
\value{
A data matrix of the same size as \code{Z} after the modified centered log-ratio transformation
}
\description{
This function transforms microbial count data using the modified centered log-ratio (mclr) transformation.
}
\details{
The centered log-ratio (clr) transformation is commonly used when working with microbiome data that is subject to total sum constraint.
In order to apply the clr transformation, one needs to have a sequence of counts that are strictly positive. When there are zeros,
a small pseudocount is necessary to ensure all values are strictly positive. However, this may unfairly bias rare variables. The modified clr (mclr)
addresses this limitation by applying the clr to positive counts only and then shifting all values to be positive
with a small constant \code{eps}. The small constant \code{eps} is defined as the absolute value of the smallest values
after clr transformation to the positive counts. For more details, see Yoon et al. (2019).
}
\examples{
set.seed(1)
x <- c(rep(0,10),rpois(10,100))
y <- c(rep(0,5),rpois(15,10))
z <- rbind(x,y)
z_mclr <- mclr(z)
# compare with clr
z_clr <- apply(z,1,compositions::clr)

}
\references{
Yoon, Grace, Irina Gaynanova, and Christian L. Müller. "Microbial networks in SPRING-Semi-parametric rank-based correlation and partial correlation estimation for quantitative microbiome data." Frontiers in Genetics 10 (2019).
}
