% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabs.R
\docType{data}
\name{metabs}
\alias{metabs}
\title{Metabolite data}
\format{A 128 by 131 data matrix of metabolite concentration profiles.}
\description{
This data matrix consists of the concentration profiles from 128 metabolites.
}
\references{
\url{https://www.nature.com/articles/srep14174}
}
\author{
Jing Ma (\email{jingma@fredhutch.org})
}
\keyword{data}
