% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{modelDensity}
\alias{modelDensity}
\title{evaluate the density in the current model}
\usage{
modelDensity(x, nu, modelFamily = "Gauss")
}
\arguments{
\item{x}{vector with points where to evaluate the density}

\item{nu}{distribution parameter}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}, \code{Poisson}}
}
\description{
evaluate the density in the current model
}
