% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTUs.R
\docType{data}
\name{OTUs}
\alias{OTUs}
\title{OTU count data}
\format{A 51 by 131 data matrix of OTU counts.}
\description{
This data matrix consists of the sequencing counts from 51 bacterial taxa.
}
\references{
\url{https://www.nature.com/articles/srep14174}
}
\author{
Jing Ma (\email{jingma@fredhutch.org})
}
\keyword{data}
