% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm_biSBM.R
\name{plotICL}
\alias{plotICL}
\title{plot ICL curve}
\usage{
plotICL(res, exclusive = TRUE)
}
\arguments{
\item{res}{output of fitNobiSBM()}

\item{exclusive}{whether the number of row clusters is equal to the number of column clusters.}
}
\value{
figure of ICL curve
}
\description{
plot ICL curve
}
\examples{
# res_gauss is the output of a call of fitNobiSBM()
plotICL(res_gauss)
}
