% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure_biSBM.R
\name{q_delta_ql_biSBM}
\alias{q_delta_ql_biSBM}
\title{auxiliary function for the computation of q-values}
\usage{
q_delta_ql_biSBM(theta, ind, t, modelFamily = "Gauss")
}
\arguments{
\item{theta}{list of parameters for a noisy bipartite stochastic block model}

\item{ind}{indicators for a pair of latent blocks}

\item{t}{l-values}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}}
}
\description{
auxiliary function for the computation of q-values
}
