% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testProcedure_biSBM.R
\name{qvaluesNobiSBM}
\alias{qvaluesNobiSBM}
\title{compute q-values in the noisy bipartite stochastic block model}
\usage{
qvaluesNobiSBM(Z1, Z2, theta, lvalues, modelFamily = "Gauss")
}
\arguments{
\item{Z1}{a node clustering for rows}

\item{Z2}{a node clustering for columns}

\item{theta}{list of parameters for a noisy bipartite stochastic block model}

\item{lvalues}{conditional l-values in the noisy bipartite stochastic block model}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}}
}
\value{
q-values in the noisy bipartite stochastic block model
}
\description{
compute q-values in the noisy bipartite stochastic block model
}
