% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_corr_test_method}
\alias{run_corr_test_method}
\title{Internal function for marginal corr_test method}
\usage{
run_corr_test_method(
  X,
  Y,
  zscore_method,
  alpha,
  sbm_model,
  sbm_params,
  nb_cores
)
}
\description{
Internal function for marginal corr_test method
}
