% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_correlation_method}
\alias{run_correlation_method}
\title{Internal function for correlation methods}
\usage{
run_correlation_method(X, Y, method, lambda)
}
\description{
Internal function for correlation methods
}
