% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_sbm_method}
\alias{run_sbm_method}
\title{Internal function for marginal SBM method}
\usage{
run_sbm_method(
  X,
  Y,
  method,
  zscore_method,
  alpha,
  sbm_model,
  sbm_params,
  init_params,
  nb_cores,
  result
)
}
\description{
Internal function for marginal SBM method
}
