% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMinitialization_biSBM.R
\name{spectralClustering}
\alias{spectralClustering}
\title{spectral clustering with absolute values for bipartite graphs}
\usage{
spectralClustering(A, K1, K2)
}
\arguments{
\item{A}{adjacency matrix}

\item{K1}{number of desired clusters over rows}

\item{K2}{number of desired clusters over columns}
}
\value{
a list containing a node clustering of rows into K1 groups and a node clustering of columns into K2 groups
}
\description{
performs absolute spectral clustering of an adjacency matrix
}
