% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{teststat_general}
\alias{teststat_general}
\title{Function to calculate test statistics}
\usage{
teststat_general(X, Y = NULL)
}
\arguments{
\item{X, Y}{feature by sample matrices}
}
\description{
Function to calculate test statistics
}
