% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions_biSBM.R
\name{teststat_general_2sample}
\alias{teststat_general_2sample}
\title{Function to calculate two-sample test statistics}
\usage{
teststat_general_2sample(sample1, sample2)
}
\description{
Function to calculate two-sample test statistics
}
