% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMalgorithm_biSBM.R
\name{VEstep}
\alias{VEstep}
\title{VE-step (updated for biSBM)}
\usage{
VEstep(VE, mstep, dataMatrix, modelFamily, fix.iter = 5)
}
\arguments{
\item{VE}{list with variational parameters beta and rho}

\item{mstep}{list with current model parameters and additional auxiliary terms}

\item{dataMatrix}{observed dense adjacency matrix}

\item{modelFamily}{probability distribution for the edges. Possible values:
\code{Gauss}, \code{Gamma}}

\item{fix.iter}{maximal number of iterations for fixed point equation}
}
\value{
updated list \code{VE} with variational parameters beta and rho
}
\description{
performs one VE-step, that is, update of beta and rho
}
